update license: Remove year; add dry-run update func
The Linux Foundation recommends not to list the attribution years at all.
https://www.linuxfoundation.org/blog/blog/copyright-notices-in-open-source-software-projects

+++ update_license_header.py
@@ -16,7 +16,7 @@ import re
 license_txt = """\
 (*)~---------------------------------------------------------------------------
 Pupil - eye tracking platform
-Copyright (C) 2012-2022 Pupil Labs
+Copyright (C) Pupil Labs
 
 Distributed under the terms of the GNU
 Lesser General Public License (LGPL v3.0).
@@ -71,7 +71,7 @@ def get_files(start_dir, includes, excludes):
     return match_files
 
 
-def write_header(file_name, license_txt, delete_header=False):
+def write_header(file_name, license_txt, delete_header=False, dry_run: bool = False):
     # find and replace license header
     # or add new header if not existing
     c_comment = ["/*\n", "\n*/\n"]
@@ -91,21 +91,25 @@ def write_header(file_name, license_txt, delete_header=False):
     except UnicodeDecodeError:
         return
 
-    with open(file_name, "w") as modified:
-        if re.findall(pattern, data):
-            if delete_header:
-                license_txt = ""
-            # if header already exists, then update, but dont add the last newline.
-            modified.write(re.sub(pattern, license_txt[:-1], data))
-        else:
-            # else write the license header
-            modified.write(license_txt + data)
+    if not dry_run:
+        with open(file_name, "w") as modified:
+            if re.findall(pattern, data):
+                if delete_header:
+                    license_txt = ""
+                # if header already exists, then update, but dont add the last newline.
+                modified.write(re.sub(pattern, license_txt, data))
+            else:
+                # else write the license header
+                modified.write(license_txt + data)
+    else:
+        print(f"Would have modified {file_name}")
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
     parser.add_argument("-d", "--directory", default=".")
     parser.add_argument("--delete", action="store_true")
+    parser.add_argument("-n", "--dry-run", action="store_true")
     args = parser.parse_args()
 
     # Add a license/docstring header to selected files
@@ -114,4 +118,4 @@ if __name__ == "__main__":
 
     for f in match_files:
         print(f"Checking {f}")
-        write_header(f, license_txt, delete_header=args.delete)
+        write_header(f, license_txt, delete_header=args.delete, dry_run=args.dry_run)
